src/requirements.txt
alembic
Babel
bcrypt
bleach
cffi
cryptography
email-validator
Flask
flask-babel
Flask-Caching
Flask-Login
Flask-Migrate
Flask-Principal
Flask-Security-Too
Flask-SocketIO
Flask-SQLAlchemy
Flask-WTF
google-analytics-data
google-api-core
google-auth
googleapis-common-protos
greenlet
grpcio
itsdangerous
Jinja2
mailjet-rest
passlib
phonenumberslite
psycopg2-binary
PyJWT
PyQRCode
python-dotenv
python-engineio
python-socketio
pytz
simple-websocket
SQLAlchemy
twilio
typing_extensions
urllib3
webencodings
Werkzeug
wsproto
WTForms
yarl
zxcvbn
infobip-api-python-sdk
flask-admin